requirements.txt
fastapi>=0.110
uvicorn>=0.30
python-multipart>=0.0.9

pydantic>=2.6
python-dotenv>=1.0
openai>=1.0
httpx>=0.28

torch>=2.2
transformers>=4.40
datasets>=2.18
scikit-learn>=1.4

pandas>=2.0
numpy>=1.24
jieba>=0.42
flashtext>=2.7
openpyxl>=3.1

tqdm>=4.66
matplotlib>=3.8

sqlalchemy>=2.0
alembic>=1.13
celery>=5.4
redis>=5.0
psycopg2-binary>=2.9

pytest>=8.2
